+++ app/src/main/java/org/mozilla/fenix/quickactionsheet/QuickActionSheetView.kt
@@ -18,24 +18,25 @@ import kotlinx.android.synthetic.main.layout_quick_action_sheet.*
 import kotlinx.android.synthetic.main.layout_quick_action_sheet.view.*
 import mozilla.components.support.ktx.android.view.putCompoundDrawablesRelativeWithIntrinsicBounds
 import org.mozilla.fenix.R
+import org.mozilla.fenix.components.toolbar.BrowserState
 import org.mozilla.fenix.utils.Settings
 
-interface QuickActionSheetInteractor {
-    fun onOpened()
-    fun onClosed()
-    fun onSharedPressed()
-    fun onDownloadsPressed()
-    fun onBookmarkPressed()
-    fun onReadPressed()
-    fun onAppearancePressed()
-    fun onOpenAppLinkPressed()
+interface QuickActionSheetViewInteractor {
+    fun onQuickActionSheetOpened()
+    fun onQuickActionSheetClosed()
+    fun onQuickActionSheetSharePressed()
+    fun onQuickActionSheetDownloadPressed()
+    fun onQuickActionSheetBookmarkPressed()
+    fun onQuickActionSheetReadPressed()
+    fun onQuickActionSheetAppearancePressed()
+    fun onQuickActionSheetOpenLinkPressed()
 }
 /**
  * View for the quick action sheet that slides out from the toolbar.
  */
-class QuickActionView(
+class QuickActionSheetView(
     override val containerView: ViewGroup,
-    private val interactor: QuickActionSheetInteractor
+    private val interactor: QuickActionSheetViewInteractor
 ) : LayoutContainer, View.OnClickListener {
 
     val view: NestedScrollView = LayoutInflater.from(containerView.context)
@@ -51,9 +52,9 @@ class QuickActionView(
                 updateImportantForAccessibility(state)
 
                 if (state == BottomSheetBehavior.STATE_EXPANDED) {
-                    interactor.onOpened()
+                    interactor.onQuickActionSheetOpened()
                 } else if (state == BottomSheetBehavior.STATE_COLLAPSED) {
-                    interactor.onClosed()
+                    interactor.onQuickActionSheetClosed()
                 }
             }
 
@@ -77,12 +78,12 @@ class QuickActionView(
      */
     override fun onClick(button: View) {
         when (button.id) {
-            R.id.quick_action_share -> interactor.onSharedPressed()
-            R.id.quick_action_downloads -> interactor.onDownloadsPressed()
-            R.id.quick_action_bookmark -> interactor.onBookmarkPressed()
-            R.id.quick_action_read -> interactor.onReadPressed()
-            R.id.quick_action_appearance -> interactor.onAppearancePressed()
-            R.id.quick_action_open_app_link -> interactor.onOpenAppLinkPressed()
+            R.id.quick_action_share -> interactor.onQuickActionSheetSharePressed()
+            R.id.quick_action_downloads -> interactor.onQuickActionSheetDownloadPressed()
+            R.id.quick_action_bookmark -> interactor.onQuickActionSheetBookmarkPressed()
+            R.id.quick_action_read -> interactor.onQuickActionSheetReadPressed()
+            R.id.quick_action_appearance -> interactor.onQuickActionSheetAppearancePressed()
+            R.id.quick_action_open_app_link -> interactor.onQuickActionSheetOpenLinkPressed()
             else -> return
         }
         quickActionSheetBehavior.state = BottomSheetBehavior.STATE_COLLAPSED
@@ -108,27 +109,32 @@ class QuickActionView(
         }
     }
 
-    fun update(state: QuickActionSheetState) {
-        view.quick_action_read.isVisible = state.readable
-        view.quick_action_read.isSelected = state.readerActive
+    fun update(state: BrowserState) {
+        val quickActionSheetState = state.quickActionSheetState
+        view.quick_action_read.isVisible = quickActionSheetState.readable
+        view.quick_action_read.isSelected = quickActionSheetState.readerActive
         view.quick_action_read.text = view.context.getString(
-            if (state.readerActive) R.string.quick_action_read_close else R.string.quick_action_read
+            if (quickActionSheetState.readerActive) R.string.quick_action_read_close else R.string.quick_action_read
         )
-        notifyReaderModeButton(state.readable)
+        notifyReaderModeButton(quickActionSheetState.readable)
 
-        view.quick_action_appearance.isVisible = state.readerActive
+        view.quick_action_appearance.isVisible = quickActionSheetState.readerActive
 
-        view.quick_action_bookmark.isSelected = state.bookmarked
+        view.quick_action_bookmark.isSelected = quickActionSheetState.bookmarked
         view.quick_action_bookmark.text = view.context.getString(
-            if (state.bookmarked) R.string.quick_action_bookmark_edit else R.string.quick_action_bookmark
+            if (quickActionSheetState.bookmarked) {
+                R.string.quick_action_bookmark_edit
+            } else {
+                R.string.quick_action_bookmark
+            }
         )
 
-        if (state.bounceNeeded && Settings.getInstance(view.context).shouldAutoBounceQuickActionSheet) {
+        if (quickActionSheetState.bounceNeeded && Settings.getInstance(view.context).shouldAutoBounceQuickActionSheet) {
             quickActionSheet.bounceSheet()
         }
 
         view.quick_action_open_app_link.apply {
-            visibility = if (state.isAppLink) View.VISIBLE else View.GONE
+            visibility = if (quickActionSheetState.isAppLink) View.VISIBLE else View.GONE
         }
     }